+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * [UIQM-195](https://issues.folio.org/browse/UIQM-195) Changes for MARC Authorities App: Closing third pane does not resize the second pane.
 * [UIQM-198](https://issues.folio.org/browse/UIQM-198) Saving a MARC holdings record upon creation/update displays Instance record then Holdings record.
 * [UIQM-199](https://issues.folio.org/browse/UIQM-199) MARC Authority: Implement App context menu and keyboard shortcuts.
+* [UIQM-196](https://issues.folio.org/browse/UIQM-196) Update error message when user adds multiple 1XXs.
 * [UIQM-190](https://issues.folio.org/browse/UIQM-190) Edit MARC holdings (quickMARC) update paneheader display.
 * [UIQM-189](https://issues.folio.org/browse/UIQM-189) Add/Edit MARC holdings record: Improve error messaging when user enters an invalid Location (852 $b) value.
 * [UIQM-197](https://issues.folio.org/browse/UIQM-197) Edit MARC authority record | Remove last 1XX crashes MARC Authority app.